FederatedAgents/requirements.txt
absl-py==0.15.0
astunparse==1.6.3
cachetools==4.2.4
certifi==2022.12.7
charset-normalizer==2.1.1
clang==5.0
cloudpickle==2.2.0
decorator==5.1.1
dm-tree==0.1.8
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
gym==0.21.0
h5py==3.1.0
idna==3.4
importlib-metadata==6.0.0
influxdb==5.3.1
keras==2.6.0
Keras-Preprocessing==1.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
msgpack==1.0.4
numpy==1.19.5
oauthlib==3.2.2
opt-einsum==3.3.0
pandas==1.1.5
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2022.7
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
Rx==3.2.0
six==1.15.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
tensorflow-probability==0.14.1
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.14
websocket-client==1.3.1
Werkzeug==2.2.2
wrapt==1.12.1
zipp==3.11.0
zmq==0.0.0
